# openapi7.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hamdaz ERP",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/labels": {
   "get": {
    "tags": [
     "labels"
    ],
    "summary": "The label catalogue",
    "operationId": "index_api_v1_labels_get",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "description": "Include this team's own labels",
       "title": "Team"
      },
      "description": "Include this team's own labels"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/LabelOut"
         },
         "title": "Response Index Api V1 Labels Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "labels"
    ],
    "summary": "Add a label",
    "operationId": "create_api_v1_labels_post",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "description": "Scope it to one team",
       "title": "Team"
      },
      "description": "Scope it to one team"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LabelIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LabelOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/labels/{key}": {
   "patch": {
    "tags": [
     "labels"
    ],
    "summary": "Rename or restyle a label",
    "description": "Change the wording, not the identity.\n\nThe ``key`` cannot be edited: the assignment policy, every assignment and\nevery stored run refer to it, and changing it would detach all of them at\nonce. This is what the delete refusal on a system label points you at.",
    "operationId": "update_api_v1_labels__key__patch",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     },
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Team"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LabelUpdateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LabelOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "labels"
    ],
    "summary": "Remove a label",
    "operationId": "remove_api_v1_labels__key__delete",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     },
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Team"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/labels/people": {
   "get": {
    "tags": [
     "labels"
    ],
    "summary": "Everyone's labels, including the automatic ones",
    "description": "Labels as they apply *today*.\n\nIncludes the two nobody assigns: someone on approved leave shows ``on-leave``\nuntil the day their leave ends, and a recent starter shows ``new-joiner``\nuntil the window passes. Neither is stored, so neither can be stale.",
    "operationId": "people_api_v1_labels_people_get",
    "parameters": [
     {
      "name": "team",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "description": "Judge team-scoped labels against this team",
       "title": "Team"
      },
      "description": "Judge team-scoped labels against this team"
     },
     {
      "name": "new_joiner_days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 3650,
       "minimum": 0,
       "default": 90,
       "title": "New Joiner Days"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PersonLabelsOut"
         },
         "title": "Response People Api V1 Labels People Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/labels/assign": {
   "post": {
    "tags": [
     "labels"
    ],
    "summary": "Give someone a label",
    "operationId": "assign_api_v1_labels_assign_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssignLabelIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PersonLabelsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/labels/unassign": {
   "post": {
    "tags": [
     "labels"
    ],
    "summary": "Take a label away",
    "operationId": "unassign_api_v1_labels_unassign_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssignLabelIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PersonLabelsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/labels/people/{user_id}/joined-on": {
   "put": {
    "tags": [
     "labels"
    ],
    "summary": "Set the date the new-joiner rule counts from",
    "description": "Entra does not give us a hire date, so this is where a real one goes.\n\nLeft unset, the new-joiner rule falls back to when the person first appeared\nin this system \u2014 usually close, occasionally very wrong for somebody who was\nhere long before the ERP was.",
    "operationId": "set_joined_on_api_v1_labels_people__user_id__joined_on_put",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JoinedOnIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PersonLabelsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AssignLabelIn": {
    "properties": {
     "user_id": {
      "type": "string",
      "format": "uuid",
      "title": "User Id"
     },
     "label_key": {
      "type": "string",
      "title": "Label Key"
     },
     "team_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team Id"
     },
     "expires_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "label_key"
    ],
    "title": "AssignLabelIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HeldLabelOut": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "kind": {
      "$ref": "#/components/schemas/LabelKind"
     },
     "source": {
      "$ref": "#/components/schemas/LabelSource"
     },
     "expires_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "key",
     "name",
     "kind",
     "source"
    ],
    "title": "HeldLabelOut",
    "description": "A label somebody actually holds right now."
   },
...
   "JoinedOnIn": {
    "properties": {
     "joined_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Joined On"
     }
    },
    "type": "object",
    "required": [
     "joined_on"
    ],
    "title": "JoinedOnIn",
    "description": "The date the new-joiner rule counts from."
   },
   "LabelIn": {
    "properties": {
     "key": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Key"
     },
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     },
     "kind": {
      "$ref": "#/components/schemas/LabelKind"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "color": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 16
       },
       {
        "type": "null"
       }
      ],
      "title": "Color"
     }
    },
    "type": "object",
    "required": [
     "key",
     "name",
     "kind"
    ],
    "title": "LabelIn"
   },
   "LabelKind": {
    "type": "string",
    "enum": [
     "category",
     "status",
     "skill"
    ],
    "title": "LabelKind"
   },
   "LabelOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "key": {
      "type": "string",
      "title": "Key"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "kind": {
      "$ref": "#/components/schemas/LabelKind"
     },
     "color": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Color"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "is_system": {
      "type": "boolean",
      "title": "Is System"
     },
     "team_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Team Id"
     },
     "derived": {
      "type": "boolean",
      "title": "Derived",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "id",
     "key",
     "name",
     "kind",
     "color",
     "description",
     "is_system",
     "team_id"
    ],
    "title": "LabelOut"
   },
   "LabelSource": {
    "type": "string",
    "enum": [
     "manual",
     "derived"
    ],
    "title": "LabelSource",
    "description": "How someone came to hold a label.\n\nWorth recording: an admin removing a label they granted is an edit, while an\nadmin removing a derived one is a request for it to come back tomorrow."
   },
   "LabelUpdateIn": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "color": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 16
       },
       {
        "type": "null"
       }
      ],
      "title": "Color"
     },
     "kind": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/LabelKind"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "title": "LabelUpdateIn",
    "description": "A patch. The ``key`` is not here on purpose \u2014 see ``update_label``."
   },
...
   "PersonLabelsOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "format": "uuid",
      "title": "User Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "joined_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Joined On"
     },
     "labels": {
      "items": {
       "$ref": "#/components/schemas/HeldLabelOut"
      },
      "type": "array",
      "title": "Labels"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "display_name",
     "email",
     "joined_on",
     "labels"
    ],
    "title": "PersonLabelsOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}